+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 #### 7.3.4-SNAPSHOT
 
 * PR [#284](https://github.com/azagniotov/stubby4j/pull/284) - During YAML parsing, checking if configured properties belong in their respective top-level objects (https://github.com/azagniotov)
+* PR [#286](https://github.com/azagniotov/stubby4j/pull/286) - Upgraded jetty to `9.4.43.v20210629` and SnakeYAML to `v1.29` (https://github.com/azagniotov)
 
 #### 7.3.3